refactor: centralize Pixoo instance in shared dependencies module
- Create src/pixoo_rest/dependencies.py with get_pixoo() dependency
- Remove duplicate _pixoo_instance from each router
- All routers now use the same global instance
- Cleaner architecture, single source of truth
- Tested: draw/pixel endpoint works correctly
