+++ src/pixoo_rest/api/send.py
@@ -1,28 +1,12 @@
-"""Send endpoints - sending text and images to the Pixoo display."""
+"""Send endpoints for the Pixoo REST API."""
 
 from fastapi import APIRouter, Depends, HTTPException
 from pixoo import Pixoo
 
+from pixoo_rest.dependencies import get_pixoo
 from pixoo_rest.models.requests import SendTextRequest, SuccessResponse
 
 router = APIRouter(prefix="/send", tags=["send"])
-
-
-# This will be injected by the app
-_pixoo_instance: Pixoo | None = None
-
-
-def set_pixoo_instance(pixoo_instance: Pixoo):
-    """Set the global Pixoo instance."""
-    global _pixoo_instance
-    _pixoo_instance = pixoo_instance
-
-
-def get_pixoo() -> Pixoo:
-    """Dependency to get the Pixoo instance."""
-    if _pixoo_instance is None:
-        raise HTTPException(status_code=503, detail="Pixoo device not initialized")
-    return _pixoo_instance
 
 
 @router.post("/text", response_model=SuccessResponse)
